connect tests
